This is a merge to make sure that everything is available to everyone

+++ src/main/java/edu/neumont/dal/SlateDAL.java
@@ -10,10 +10,10 @@ public interface SlateDAL {
 	public abstract boolean createSlate(String name, String description,
 			LocalDateTime dueDate);
 
-	public abstract Slate retrieveSlate(int index);
+	public abstract Slate retrieveSlate(long index);
 
-	public abstract void updateSlate(int index, Slate updatedSlate);
+	public abstract void updateSlate(long index, Slate updatedSlate);
 	
-	public abstract void deleteSlate(int index);
+	public abstract void deleteSlate(long index);
 
 }
